requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
discord.py==2.0.1
frozenlist==1.3.1
multidict==6.0.2
PyNaCl==1.5.0
python-dotenv==0.21.0
yarl==1.8.1
discord.py[voice]==2.0.1
pymongo==4.3.2
